Move record::c_api back in the main record module.
After `record.rs` was split into multiple modules, the C API disappeared
from the compiled .so library.  This is likely due to this compiler bug:
https://github.com/rust-lang/rust/issues/50007

As a workaround until the bug is fixed, we move the API from `record/c_api.rs`
to `record/mod.rs`.
